requirements.txt
numpy==1.22.3
matplotlib==3.5.1
pyyaml==6.0
